{"schema":"decodex/commit/1","summary":"Split live adapter materialization model modules","authority":"manual"}
